requirements.txt
torch==2.4.1
torchvision==0.19.1
pyyaml==6.0.2
Pillow==10.4.0
ultralytics==8.3.13
